Move `bufferToBlob`
