+++ packages/media-utils/README.md
@@ -6,18 +6,6 @@ A set of media-related helper utilities.
 
 <!-- START TOKEN(Autogenerated API docs) -->
 
-### bufferToBlob
-
-Creates a Blob from a given ArrayBuffer.
-
-_Parameters_
-
--   _buffer_ `ArrayBuffer`: ArrayBuffer instance.
--   _width_ `number`: Desired width.
--   _height_ `number`: Desired height.
--   _type_ `'image/jpeg' | 'image/png' | 'image/webp'`: Desired mime type.
--   _quality_ Desired image quality.
-
 ### cloneFile
 
 Clones a given file object.
