+++ packages/media-utils/src/index.ts
@@ -176,38 +176,3 @@ export function getCanvasBlob(
 		);
 	} );
 }
-
-/**
- * Creates a Blob from a given ArrayBuffer.
- *
- * @param buffer  ArrayBuffer instance.
- * @param width   Desired width.
- * @param height  Desired height.
- * @param type    Desired mime type.
- * @param quality Desired image quality.
- */
-export async function bufferToBlob(
-	buffer: ArrayBuffer,
-	width: number,
-	height: number,
-	type: 'image/jpeg' | 'image/png' | 'image/webp' = 'image/jpeg',
-	quality = 0.82
-): Promise< Blob > {
-	const canvas = new OffscreenCanvas( width, height );
-
-	const ctx = canvas.getContext( '2d' );
-
-	if ( ! ctx ) {
-		throw new Error( 'Could not get context' );
-	}
-
-	const imageData = new ImageData(
-		new Uint8ClampedArray( buffer ),
-		width,
-		height
-	);
-
-	ctx.putImageData( imageData, 0, 0 );
-
-	return canvas.convertToBlob( { type, quality } );
-}
